+++ packages/react-components/source/react/library/sidebar/SidebarHeader.js
@@ -5,50 +5,51 @@ import Logo from '../logo';
 const propTypes = {
   ariaLabel: PropTypes.string,
   as: PropTypes.elementType,
+  children: PropTypes.node,
+  className: PropTypes.string,
   logo: PropTypes.string,
   minimized: PropTypes.bool,
   onClick: PropTypes.func,
 };
-
 const defaultProps = {
   ariaLabel: 'Return to home',
   as: 'button',
+  children: null,
+  className: '',
   logo: '',
   minimized: false,
   onClick() {},
 };
 
-class SidebarHeader extends React.Component {
-  renderLogo() {
-    const { logo, minimized } = this.props;
-    let jsx;
+const SidebarHeader = ({
+  ariaLabel,
+  as,
+  children,
+  className,
+  logo,
+  minimized,
+  ...rest
+}) => {
+  const Component = as;
 
-    if (logo) {
-      jsx = (
+  return (
+    <Component
+      className={`rc-sidebar-header ${className}`}
+      {...rest}
+      aria-label={ariaLabel}
+    >
+      {logo && (
         <Logo
           inverted
           expanded
           product={logo}
           type={minimized ? 'bug' : 'full'}
         />
-      );
-    }
-
-    return jsx;
-  }
-
-  render() {
-    const { as, minimized, ariaLabel, ...rest } = this.props;
-    const logo = this.renderLogo();
-    const Component = as;
-
-    return (
-      <Component className="rc-sidebar-header" {...rest} aria-label={ariaLabel}>
-        {logo}
-      </Component>
-    );
-  }
-}
+      )}
+      {children}
+    </Component>
+  );
+};
 
 SidebarHeader.propTypes = propTypes;
 SidebarHeader.defaultProps = defaultProps;